fix reference DESCR length for YiiP test
The YiiP test failed locally for me. I adjusted the reference number of characters in the
DESCR file (same as with wc) but because these tests are NOT run in CI, I am not 100%
sure if this is correct or a local issue.

I added comments to test_datasets.py about limitations and how to execute locally.